Remove 2 pointless fade effects and drawing outside widget.
In gmpc-metadata-image add a function that creates it for a given size.
withouth size should not be used anymore. (if size it set before
realizing an out-of-widget draw can occure.).

In gmpc-image.vala don't fade in when there is no image yet.

metadata browser, use the new create_With_size.